backend/requirements.txt

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0

grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

pika>=1.3.0

python-multipart>=0.0.6

cryptography>=42.0.0

pytest>=7.0.0
httpx>=0.25.0
